requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
duckduckgo-search
sentence-transformers
faiss-cpu
numpy
# 桌面客户端依赖（client_app.py）
pywebview>=4.0
# 后端拼音引擎
pypinyin
jieba
